Ingles/Fast---Ingles/backend/requirements.txt

# Fast-Ingles Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
supabase>=2.0,<3.0
firebase-admin==6.6.0  # Added for Firebase Auth

# AI Providers
google-generativeai==0.8.3
anthropic==0.40.0
openai==1.58.1

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
python-multipart==0.0.18

# MinIO Storage
minio==7.2.12

# Development
pytest==8.3.4
pytest-asyncio==0.25.0

# TTS
gTTS==2.5.4
